requirements.txt
streamlit>=1.30.0
PyGithub>=2.1.1
pydantic>=2.0.0
pydantic-ai>=0.0.31